Interfaces Hierarchy simplification
All interfaces are defined under one subnamespace
