+++ src/Components/Port.php
@@ -11,7 +11,7 @@
  */
 namespace League\Uri\Components;
 
-use League\Uri\Interfaces\Components\Port as PortInterface;
+use League\Uri\Interfaces\Port as PortInterface;
 
 /**
  * Value object representing a URI port component.
