+++ src/Components/Scheme.php
@@ -12,7 +12,7 @@
 namespace League\Uri\Components;
 
 use InvalidArgumentException;
-use League\Uri\Interfaces\Components\Scheme as SchemeInterface;
+use League\Uri\Interfaces\Scheme as SchemeInterface;
 
 /**
  * Value object representing a URI scheme component.
