+++ src/Interfaces/Port.php
@@ -9,10 +9,10 @@
  * @version   4.0.0
  * @link      https://github.com/thephpleague/uri/
  */
-namespace League\Uri\Interfaces\Components;
+namespace League\Uri\Interfaces;
 
 /**
- * Value object representing a URI component.
+ * Value object representing a URI Port component.
  *
  * Instances of this interface are considered immutable; all methods that
  * might change state MUST be implemented such that they retain the internal
@@ -22,18 +22,28 @@ namespace League\Uri\Interfaces\Components;
  * @package League.uri
  * @author  Ignace Nyamagana Butera <nyamsprod@gmail.com>
  * @since   4.0.0
+ * @see     https://tools.ietf.org/html/rfc3986#section-3.2.3
  */
-interface Component extends UriPart
+interface Port extends Component
 {
     /**
-     * Returns an instance with the specified string
+     * Maximum port number.
      *
-     * This method MUST retain the state of the current instance, and return
-     * an instance that contains the modified data
-     *
-     * @param string $value
-     *
-     * @return static
+     * @var int
      */
-    public function modify($value);
+    const MAXIMUM = 65535;
+
+    /**
+     * Minimum port number.
+     *
+     * @var int
+     */
+    const MINIMUM = 1;
+
+    /**
+     * Return an integer representation of the Port component
+     *
+     * @return null|int
+     */
+    public function toInt();
 }
