+++ src/Modifiers/MergeQuery.php
@@ -11,7 +11,7 @@
  */
 namespace League\Uri\Modifiers;
 
-use League\Uri\Interfaces\Components\Query;
+use League\Uri\Interfaces\Query;
 use League\Uri\Modifiers\Filters\QueryString;
 
 /**
