+++ src/Modifiers/PrependSegment.php
@@ -11,7 +11,7 @@
  */
 namespace League\Uri\Modifiers;
 
-use League\Uri\Interfaces\Components\Path;
+use League\Uri\Interfaces\Path;
 use League\Uri\Modifiers\Filters\Segment;
 
 /**
