+++ src/Modifiers/ReplaceLabel.php
@@ -11,7 +11,7 @@
  */
 namespace League\Uri\Modifiers;
 
-use League\Uri\Interfaces\Components\Host;
+use League\Uri\Interfaces\Host;
 use League\Uri\Modifiers\Filters\Label;
 use League\Uri\Modifiers\Filters\Offset;
