+++ src/Types/ImmutableCollectionTrait.php
@@ -13,7 +13,7 @@ namespace League\Uri\Types;
 
 use ArrayIterator;
 use InvalidArgumentException;
-use League\Uri\Interfaces\Components\Collection;
+use League\Uri\Interfaces\Collection;
 use LogicException;
 use Traversable;
