+++ test/Components/QueryTest.php
@@ -88,7 +88,7 @@ class QueryTest extends PHPUnit_Framework_TestCase
         $empty_query = new Query();
         $this->assertFalse($empty_query->sameValueAs($this->query));
         $query = $empty_query->merge($this->query);
-        $this->assertInstanceof('League\Uri\Interfaces\Components\Query', $query);
+        $this->assertInstanceof('League\Uri\Interfaces\Query', $query);
         $this->assertTrue($query->sameValueAs($this->query));
     }
